Update (2023.07.12)
31170: Fix conflicting types for MIN2()
31168: LA port of 8303040: linux PPC64le: Implementation of Foreign Function & Memory API (Preview)
31096: LA port of 8306302: C2 Superword fix: use VectorMaskCmp and VectorBlend instead of CMoveVF/D

+++ src/hotspot/cpu/loongarch/sharedRuntime_loongarch_64.cpp
@@ -929,7 +929,7 @@ static void verify_oop_args(MacroAssembler* masm,
                             const VMRegPair* regs) {
   if (VerifyOops) {
     // verify too many args may overflow the code buffer
-    int arg_size = MIN2(64, method->size_of_parameters());
+    int arg_size = MIN2(64, (int)(method->size_of_parameters()));
 
     for (int i = 0; i < arg_size; i++) {
       if (is_reference_type(sig_bt[i])) {
